Rename source files to kebab case and update references in README
