+++ gulp/tasks/banners.js
@@ -3,7 +3,7 @@
  */
 import gulp from 'gulp';
 import gulpBanner from 'gulp-banner';
-import {isProdBuild} from '../commandLineArgs';
+import {isProdBuild} from '../command-line-args';
 import {generateBanners, pkgJson, settings} from '../config';
 import merge from 'merge-stream';
