+++ gulp/tasks/commit-changes.js
